feat: Add routing and theming to App component
This commit adds routing to the App component using react-router-dom. It also adds theming using @mui/material and the createTheme function. The theme is based on the mode state that is obtained using useSelector from the Redux store. The app's root component is also wrapped in the ThemeProvider component, which sets the theme for all child components. The CssBaseline component is added to provide global styling to the app.