Add Friend component
displays a user's friend with their name, subtitle, and profile picture. The component also has a button that allows the user to add or remove the friend from their friends list. The friend's information is obtained from the store, and the patchFriend function sends a PATCH request to update the user's friends list. The component also navigates to the friend's profile page when clicked.