Feat: add MyPostWidget Component
This commit adds a new component called MyPostWidget which allows users to create a new post by entering text and optionally attaching an image. The component also includes functionality to upload the post to the server and update the state of the application with the newly created post. The component includes a number of Material UI icons and styles to provide a user-friendly interface.